refactor: rename and split storage across different files
